+++ CHANGELOG.md
@@ -10,6 +10,14 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 ### Fixed
 ### Changed
 
+## [0.2.0] - 2017-02-09
+
+### Breaking Changes
+- check-topics.rb has been removed
+
+### Added
+- Add check-topic.rb
+
 ## [0.1.1] - 2017-02-04
 
 ### Fixed